requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
decorator==4.0.4
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-admin-bootstrapped==2.5.6
django-allauth==0.23.0
django-authtools==1.2.0
django-braces==1.8.1
django-celery==3.1.17
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-discover-runner==1.0
django-environ==0.4.0
django-filter==0.11.0
django-oauth-toolkit==0.9.0
django-rest-auth==0.5.0
django-rest-framework-social-oauth2==1.0.2
django-rest-swagger==0.3.4
django-role-permissions==0.6.2
django-toolbelt==0.0.1
djangorestframework==3.3.0
easy-thumbnails==2.2
gunicorn==19.3.0
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
kombu==3.0.29
Markdown==2.6.2
oauthlib==1.0.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.6.1
ptyprocess==0.5
Pygments==2.0.2
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.13
pytz==2015.7
PyYAML==3.11
redis==2.10.3
requests==2.8.1
requests-oauthlib==0.5.0
rest-social-auth==0.2.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.16
static3==0.6.1
traitlets==4.0.0
wheel==0.24.0
